+++ udata/core/contact_point/models.py
@@ -5,7 +5,15 @@ __all__ = ("ContactPoint",)
 
 
 class ContactPoint(db.Document, Owned):
-    email = db.StringField(max_length=255, required=True)
     name = db.StringField(max_length=255, required=True)
+    email = db.StringField(max_length=255)
+    contact_form = db.URLField()
 
     meta = {"queryset_class": OwnedQuerySet}
+
+    def validate(self, clean=True):
+        if not self.email and not self.contact_form:
+            raise db.ValidationError(
+                "At least an email or a contact form is required for a contact point"
+            )
+        return super().validate(clean=clean)